glesys-go v3
* Include the new Float for cost.
* move v3 to root of project
