--- v2/go.mod
@@ -1,12 +1,0 @@
-module github.com/glesys/glesys-go/v2
-
-require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/imdario/mergo v0.3.7
-	github.com/kr/pretty v0.2.0 // indirect
-	github.com/stretchr/testify v1.3.0
-	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
-	gopkg.in/yaml.v2 v2.2.2 // indirect
-)
-
-go 1.13
